+++ src/components/beta/layout/side_menu.tsx
@@ -81,13 +81,13 @@ const MENU_CONFIG: TMenuOption[] = [
           {
             type: SubMenuOptionType.LINK,
             title: 'INPUT_INVOICE',
-            link: ISUNFA_ROUTE.INPUT_CERTIFICATE_LIST,
+            link: ISUNFA_ROUTE.INPUT_INVOICE_LIST,
             needToConnectAccountBook: true,
           },
           {
             type: SubMenuOptionType.LINK,
             title: 'OUTPUT_INVOICE',
-            link: ISUNFA_ROUTE.OUTPUT_CERTIFICATE_LIST,
+            link: ISUNFA_ROUTE.OUTPUT_INVOICE_LIST,
             needToConnectAccountBook: true,
           },
         ],
