+++ src/components/certificate/output_invoice_item.tsx
@@ -6,14 +6,11 @@ import Image from 'next/image';
 import Link from 'next/link';
 import { CurrencyType } from '@/constants/currency';
 import { numberWithCommas } from '@/lib/utils/common';
-import { FaCheck } from 'react-icons/fa6';
-import { RxCross2 } from 'react-icons/rx';
-import { ICertificateRC2InputUI } from '@/interfaces/certificate_rc2';
-import { DeductionType } from '@/constants/certificate';
+import { ICertificateRC2OutputUI } from '@/interfaces/certificate_rc2';
 
-interface InputCertificateListIrops {
+interface OutputInvoiceListIrops {
   activeSelection: boolean;
-  certificate: ICertificateRC2InputUI;
+  certificate: ICertificateRC2OutputUI;
   currencyAlias: CurrencyType;
   handleSelect: (ids: number[], isSelected: boolean) => void;
   onEdit: (id: number) => void;
@@ -35,7 +32,7 @@ const BorderCell: React.FC<{ isSelected: boolean; children: ReactElement; classN
   );
 };
 
-const InputCertificateItem: React.FC<InputCertificateListIrops> = ({
+const OutputInvoiceItem: React.FC<OutputInvoiceListIrops> = ({
   activeSelection,
   certificate,
   currencyAlias,
@@ -43,10 +40,6 @@ const InputCertificateItem: React.FC<InputCertificateListIrops> = ({
   onEdit,
 }) => {
   const { t } = useTranslation(['common', 'certificate', 'filter_section_type']);
-
-  const isDeductible =
-    certificate.deductionType === DeductionType.DEDUCTIBLE_PURCHASE_AND_EXPENSE ||
-    certificate.deductionType === DeductionType.DEDUCTIBLE_FIXED_ASSETS;
 
   return (
     <div
@@ -83,62 +76,32 @@ const InputCertificateItem: React.FC<InputCertificateListIrops> = ({
       <BorderCell isSelected={certificate.isSelected} className="flex w-120px gap-1">
         <div className="flex items-center gap-2">
           {certificate.incomplete && (
-            <Image
-              src="/icons/hint.svg"
-              alt="Hint"
-              width={16}
-              height={16}
-              className="download-hidden min-w-16px"
-            />
+            <Image src="/icons/hint.svg" alt="Hint" width={16} height={16} className="min-w-16px" />
           )}
           <div className="flex flex-col">
-            <div className="download-pb-4 text-text-neutral-primary">{certificate.no ?? ''}</div>
+            <div className="text-text-neutral-primary">{certificate.no ?? ''}</div>
           </div>
         </div>
       </BorderCell>
       <BorderCell isSelected={certificate.isSelected} className="row-span-full min-w-100px">
-        <div className="hide-scrollbar download-pb-4 max-h-72px w-full overflow-y-auto text-left text-text-neutral-primary">
+        <div className="hide-scrollbar max-h-72px w-full overflow-y-auto text-left text-text-neutral-primary">
           {certificate.type ? t(`filter_section_type:FILTER_SECTION_TYPE.${certificate.type}`) : ''}
-        </div>
-      </BorderCell>
-      <BorderCell isSelected={certificate.isSelected} className="row-span-full min-w-100px">
-        <div className="hide-scrollbar download-pb-4 max-h-72px w-full overflow-y-auto text-left text-text-neutral-primary">
-          {/* Info: (20250421 - Anna) deductionType */}
-          {isDeductible ? (
-            <div className="flex items-center gap-2">
-              <FaCheck className="h-6 w-6 text-green-500" />
-              <span className="text-sm text-neutral-300">
-                {certificate.deductionType === DeductionType.DEDUCTIBLE_FIXED_ASSETS
-                  ? t(`certificate:TABLE.DEDUCTIBLE_FIXED_ASSETS`)
-                  : t(`certificate:TABLE.DEDUCTIBLE_PURCHASE_AND_EXPENSE`)}
-              </span>
-            </div>
-          ) : (
-            <div className="flex items-center gap-2">
-              <RxCross2 className="h-6 w-6 text-navy-blue-400" />
-              <span className="text-sm text-neutral-300">
-                {certificate.deductionType === DeductionType.NON_DEDUCTIBLE_FIXED_ASSETS
-                  ? t(`certificate:TABLE.NON_DEDUCTIBLE_FIXED_ASSETS`)
-                  : t(`certificate:TABLE.NON_DEDUCTIBLE_PURCHASE_AND_EXPENSE`)}
-              </span>
-            </div>
-          )}
         </div>
       </BorderCell>
       <BorderCell isSelected={certificate.isSelected} className="w-100px">
         <div
-          className={`download-pb-4 w-full ${certificate.taxRate !== undefined ? 'text-left' : 'text-center'} text-text-neutral-primary`}
+          className={`w-full ${certificate.taxRate !== undefined ? 'text-left' : 'text-center'} text-text-neutral-primary`}
         >
           {certificate.taxRate !== undefined ? `Taxable ${certificate.taxRate} %` : '-'}
         </div>
       </BorderCell>
       <BorderCell isSelected={certificate.isSelected} className="row-span-full min-w-100px">
-        <div className="download-pb-4 flex flex-col items-center gap-2">
+        <div className="flex flex-col items-center gap-2">
           <div className="w-full text-left text-text-neutral-tertiary">
-            {certificate.buyerIdNumber ?? ''}
+            {certificate.salesIdNumber ?? ''}
           </div>
           <div className="w-full text-left text-text-neutral-primary">
-            {certificate.buyerName ?? ''}
+            {certificate.salesName ?? ''}
           </div>
         </div>
       </BorderCell>
@@ -150,7 +113,7 @@ const InputCertificateItem: React.FC<InputCertificateListIrops> = ({
               <div
                 className={`m-1 inline-block h-6px w-6px rounded-full bg-surface-support-strong-rose`}
               ></div>
-              <div className="download-pb-3">Pre-Tax</div>
+              <div>Pre-Tax</div>
             </div>
             <div className="text-text-neutral-primary">
               {numberWithCommas(certificate.netAmount ?? 0)}
@@ -164,7 +127,7 @@ const InputCertificateItem: React.FC<InputCertificateListIrops> = ({
               <div
                 className={`m-1 inline-block h-6px w-6px rounded-full bg-surface-support-strong-baby`}
               ></div>
-              <div className="download-pb-3">After-Tax</div>
+              <div>After-Tax</div>
             </div>
             <div className="text-text-neutral-primary">
               {numberWithCommas(certificate.totalAmount ?? 0)}
@@ -189,7 +152,7 @@ const InputCertificateItem: React.FC<InputCertificateListIrops> = ({
             </Link>
           )}
           <div className="flex items-center gap-2 text-right text-text-neutral-primary">
-            {certificate?.file.url ? (
+            {certificate?.file?.url ? (
               <Image
                 src={certificate.file.url}
                 alt="avatar"
@@ -202,7 +165,7 @@ const InputCertificateItem: React.FC<InputCertificateListIrops> = ({
                 {certificate.uploaderName.slice(0, 2).toUpperCase()}
               </span>
             )}
-            <span className="download-pb-4">{certificate.uploaderName ?? ''}</span>
+            <span>{certificate.uploaderName ?? ''}</span>
           </div>
         </div>
       </BorderCell>
@@ -210,4 +173,4 @@ const InputCertificateItem: React.FC<InputCertificateListIrops> = ({
   );
 };
 
-export default InputCertificateItem;
+export default OutputInvoiceItem;
